Set a single value to several places
Prior this change items are set to html elements with id of the same
name. Now OOMDisplay._set_item() searches for elements with same-named
id or with item name listed in class attribute.

This allows to set one value to several places to once.

Additional suffixes for pages and kbytes are added in the singular or
plural automatically.