Fix Node.js warning about wrong module type
# rollup --config rollup.config.js
(node:2833) Warning: To load an ES module, set "type": "module" in the
package.json or use the .mjs extension.
(Use `node --trace-warnings ...` to show where the warning was created)
[!] RollupError: Node tried to load your configuration file as CommonJS
even though it is likely an ES module. To resolve this, change the
extension of your configuration to ".mjs", set "type": "module" in your
package.json file or pass the "--bundleConfigAsCjs" flag.